requirements.txt
asn1crypto==1.5.1
attrs==22.1.0
callable==0.1.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
elastic-enterprise-search==8.4.0
elastic-transport==8.4.0
elasticsearch==8.8.0
filelock==3.12.2
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.0
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
oscrypto==1.3.0
packaging==23.1
protobuf==3.20.1
pycparser==2.21
pycryptodomex==3.18.0
PyJWT==2.5.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
snowflake-connector-python==3.0.4
sortedcontainers==2.4.0
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==1.26.12
Werkzeug==2.2.2
